Create separate file for ItemsListNode.
