+++ claat/nodes/node.go
@@ -120,45 +120,3 @@ func (b *node) MutateEnv(e []string) {
 	copy(b.env, e)
 	sort.Strings(b.env)
 }
-
-// TODO test ItemsList related code
-// NewItemsListNode creates a new ItemsListNode of type NodeItemsList,
-// which defaults to an unordered list.
-// Provide a positive start to make this a numbered list.
-// NodeItemsCheck and NodeItemsFAQ are always unnumbered.
-func NewItemsListNode(typ string, start int) *ItemsListNode {
-	iln := ItemsListNode{
-		node: node{typ: NodeItemsList},
-		// TODO document this
-		ListType: typ,
-		Start:    start,
-	}
-	iln.MutateBlock(true)
-	return &iln
-}
-
-// ItemsListNode containts sets of ListNode.
-// Non-zero ListType indicates an ordered list.
-type ItemsListNode struct {
-	node
-	ListType string
-	Start    int
-	Items    []*ListNode
-}
-
-// Empty returns true if every item has empty content.
-func (il *ItemsListNode) Empty() bool {
-	for _, i := range il.Items {
-		if !i.Empty() {
-			return false
-		}
-	}
-	return true
-}
-
-// NewItem creates a new ListNode and adds it to il.Items.
-func (il *ItemsListNode) NewItem(nodes ...Node) *ListNode {
-	n := NewListNode(nodes...)
-	il.Items = append(il.Items, n)
-	return n
-}
